requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Embeddings
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.16

# Vector Store
chromadb==0.4.22

# Document Parsing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.3

# Utilities
tiktoken==0.5.2
python-dotenv==1.0.1
sse-starlette==2.0.0
